+++ crypto/math/src/fft/mod.rs
@@ -5,9 +5,5 @@ pub mod errors;
 #[cfg(feature = "alloc")]
 pub mod roots_of_unity;
 
-/// Reference radix-2 FFT, used only to cross-check the production Bowers FFT.
-#[cfg(test)]
-pub(crate) mod reference_fft;
-
 #[cfg(all(test, feature = "alloc"))]
 pub(crate) mod test_helpers;
