--- crypto/math/src/fft/reference_fft.rs
@@ -1,156 +1,0 @@
-#[cfg(test)]
-use crate::field::{
-    element::FieldElement,
-    traits::{IsFFTField, IsField, IsSubFieldOf},
-};
-
-/// In-Place Radix-2 NR DIT FFT algorithm over a slice of two-adic field elements.
-/// It's required that the twiddle factors are in bit-reverse order. Else this function will not
-/// return fourier transformed values.
-/// Also the input size needs to be a power of two.
-/// It's recommended to use the current safe abstractions instead of this function.
-///
-/// Performs a fast fourier transform with the next attributes:
-/// - In-Place: an auxiliary vector of data isn't needed for the algorithm.
-/// - Radix-2: the algorithm halves the problem size log(n) times.
-/// - NR: natural to reverse order, meaning that the input is naturally ordered and the output will
-///   be bit-reversed ordered.
-/// - DIT: decimation in time
-///
-/// It supports values in a field E and domain in a subfield F.
-#[cfg(test)]
-pub fn in_place_nr_2radix_fft<F, E>(input: &mut [FieldElement<E>], twiddles: &[FieldElement<F>])
-where
-    F: IsFFTField + IsSubFieldOf<E>,
-    E: IsField,
-{
-    // divide input in groups, starting with 1, duplicating the number of groups in each stage.
-    let mut group_count = 1;
-    let mut group_size = input.len();
-
-    // for each group, there'll be group_size / 2 butterflies.
-    // a butterfly is the atomic operation of a FFT, e.g: (a, b) = (a + wb, a - wb).
-    // The 0.5 factor is what gives FFT its performance, it recursively halves the problem size
-    // (group size).
-
-    while group_count < input.len() {
-        #[allow(clippy::needless_range_loop)] // the suggestion would obfuscate a bit the algorithm
-        for group in 0..group_count {
-            let first_in_group = group * group_size;
-            let first_in_next_group = first_in_group + group_size / 2;
-
-            let w = &twiddles[group]; // a twiddle factor is used per group
-
-            for i in first_in_group..first_in_next_group {
-                let wi = w * &input[i + group_size / 2];
-
-                let y0 = &input[i] + &wi;
-                let y1 = &input[i] - &wi;
-
-                input[i] = y0;
-                input[i + group_size / 2] = y1;
-            }
-        }
-        group_count *= 2;
-        group_size /= 2;
-    }
-}
-
-/// In-Place Radix-2 RN DIT FFT algorithm. Only used in tests.
-#[cfg(test)]
-pub fn in_place_rn_2radix_fft<F>(input: &mut [FieldElement<F>], twiddles: &[FieldElement<F>])
-where
-    F: IsFFTField,
-{
-    // divide input in groups, starting with 1, duplicating the number of groups in each stage.
-    let mut group_count = 1;
-    let mut group_size = input.len();
-
-    // for each group, there'll be group_size / 2 butterflies.
-    // a butterfly is the atomic operation of a FFT, e.g: (a, b) = (a + wb, a - wb).
-    // The 0.5 factor is what gives FFT its performance, it recursively halves the problem size
-    // (group size).
-
-    while group_count < input.len() {
-        let step_to_next = 2 * group_count; // next butterfly in the group
-        let step_to_last = step_to_next * (group_size / 2 - 1);
-
-        for group in 0..group_count {
-            let w = &twiddles[group * group_size / 2];
-
-            for i in (group..=group + step_to_last).step_by(step_to_next) {
-                let wi = w * &input[i + group_count];
-
-                let y0 = &input[i] + &wi;
-                let y1 = &input[i] - &wi;
-
-                input[i] = y0;
-                input[i + group_count] = y1;
-            }
-        }
-        group_count *= 2;
-        group_size /= 2;
-    }
-}
-
-#[cfg(all(test, feature = "alloc"))]
-mod tests {
-    use crate::fft::bit_reversing::in_place_bit_reverse_permute;
-    use crate::fft::roots_of_unity::get_powers_of_primitive_root;
-    use crate::fft::test_helpers::naive_matrix_dft_test;
-    use crate::field::{test_fields::u64_test_field::U64TestField, traits::RootsConfig};
-    use proptest::{collection, prelude::*};
-
-    use super::*;
-
-    type F = U64TestField;
-    type FE = FieldElement<F>;
-
-    prop_compose! {
-        fn powers_of_two(max_exp: u8)(exp in 1..max_exp) -> usize { 1 << exp }
-        // max_exp cannot be multiple of the bits that represent a usize, generally 64 or 32.
-        // also it can't exceed the test field's two-adicity.
-    }
-    prop_compose! {
-        fn field_element()(num in any::<u64>().prop_filter("Avoid null coefficients", |x| x != &0)) -> FE {
-            FE::from(num)
-        }
-    }
-    prop_compose! {
-        fn field_vec(max_exp: u8)(vec in (1..max_exp).prop_flat_map(|i| collection::vec(field_element(), 1 << i))) -> alloc::vec::Vec<FE> {
-            vec
-        }
-    }
-    proptest! {
-        // Property-based test that ensures NR Radix-2 FFT gives the same result as a naive DFT.
-        #[test]
-        fn test_nr_2radix_fft_matches_naive_eval(coeffs in field_vec(8)) {
-            let expected = naive_matrix_dft_test(&coeffs);
-
-            let order = coeffs.len().trailing_zeros();
-            let twiddles = get_powers_of_primitive_root::<F>(order.into(), (1 << order) / 2, RootsConfig::BitReverse).unwrap();
-
-            let mut result = coeffs;
-            in_place_nr_2radix_fft::<F, F>(&mut result, &twiddles);
-            in_place_bit_reverse_permute(&mut result);
-
-            prop_assert_eq!(expected, result);
-        }
-
-        // Property-based test that ensures RN Radix-2 FFT gives the same result as a naive DFT.
-        #[test]
-        fn test_rn_2radix_fft_matches_naive_eval(coeffs in field_vec(8)) {
-            let expected = naive_matrix_dft_test(&coeffs);
-
-            let order = coeffs.len().trailing_zeros();
-            let twiddles = get_powers_of_primitive_root::<F>(order.into(), (1 << order) / 2, RootsConfig::Natural).unwrap();
-
-            let mut result = coeffs;
-            in_place_bit_reverse_permute(&mut result);
-            in_place_rn_2radix_fft(&mut result, &twiddles);
-
-            prop_assert_eq!(result, expected);
-        }
-
-    }
-}
