+++ crypto/math/src/field/extensions_goldilocks.rs
@@ -95,7 +95,7 @@ impl ByteConversion for [FpE; 3] {
 // This means Fp2 = Fp[x] / (x^2 - 7)
 // Elements are represented as a0 + a1*w where w^2 = 7
 
-type FpE = FieldElement<GoldilocksField>;
+pub(crate) type FpE = FieldElement<GoldilocksField>;
 
 /// Degree 2 extension field of Goldilocks
 #[derive(Copy, Clone, Debug)]
@@ -568,28 +568,4 @@ impl HasDefaultTranscript for Degree3GoldilocksExtensionField {
 #[inline(always)]
 fn mul_by_7(a: &FpE) -> FpE {
     FpE::from_raw(mul_by_7_raw(*a.value()))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::traits::ByteConversion;
-
-    #[test]
-    fn write_bytes_be_matches_as_bytes() {
-        let cases = [
-            FieldElement::<Degree3GoldilocksExtensionField>::zero(),
-            FieldElement::<Degree3GoldilocksExtensionField>::one(),
-            FieldElement::<Degree3GoldilocksExtensionField>::new([
-                FpE::from(1u64),
-                FpE::from(2u64),
-                FpE::from(3u64),
-            ]),
-        ];
-        for elem in &cases {
-            let mut buf = [0u8; 24];
-            elem.write_bytes_be(&mut buf);
-            assert_eq!(&buf[..], elem.as_bytes().as_slice());
-        }
-    }
 }
