+++ crypto/math/src/tests/mod.rs
@@ -1,8 +1,11 @@
 pub mod barycentric_tests;
+pub mod bit_reversing_tests;
 pub mod bowers_fft_tests;
+pub mod extensions_goldilocks_tests;
 pub mod fft_friendly_extensions_goldilocks_tests;
 pub mod fft_friendly_u64_goldilocks_tests;
 pub mod fft_tests;
 pub mod field_element_tests;
+pub mod goldilocks_tests;
 pub mod polynomial_tests;
 pub mod test_fields_tests;
